requirements.txt
Flask
openai
python-dotenv
mantium-client